Move ResultRecorder to separated file
